Support alternative Context Broker implementations
Refine text to reflect that multiple alternative Context Broker implementations may become part of the FIWARE Catalogue and any can be the basis of a "Powered by FIWARE" architecture.  Also the role of the NGSI API (currently NGSIv2, NGSI-LD in the near future) as the API enabling integration of FIWARE components.